Fixes parameter detection on bytecode->goto translation
The JVM spec document (Sec. 3.6) mandates that Java method arguments occupy the
lowest local variables. The only sound way to detect which variable slots are
function parameters is by counting the number and type of arguments in the
method descriptor.

This commit adds a new function `java_method_parameter_slots` which can
determine the number of variable slots used by a method's parameters. It also
fixes `setup_local_variables` (which were previously checking whether
`v.start_pc=0` to determine if an LVT entry is a parameter).