Don't require pointers point to classes
Previously we had an invariant that all pointers were pointers to
classes. This invariant is incorrect in at least two places: strings and
exceptions. Instead, since if it isn't a pointer to a class, it
certainly isn't a pointer to an abstract class, so we can just return
false in the event it isn't a pointer to a class.